+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 ### Snowpark pandas API Updates
 #### New Features
 - Added support for `Index.is_unique` and `Index.has_duplicates`.
+- Added support for `Index.equals`.
 
 #### New Features
 - Added support for `DataFrame.backfill`, `DataFrame.bfill`, `Series.backfill`, and `Series.bfill`.